backend/requirements.txt
websockets
python-can
bitstruct
